refactor(html): own visible-text extraction on HtmlNavigator
TextExtractor lives under HtmlNavigator so DOM chrome has one home;
HtmlExtractor keeps a constant alias for compatibility and nested
callers use Navigator directly. UTILITY_LANDMARK_TAGS becomes a Set
for hot-path ancestor membership checks.
