+++ lib/html2rss/html_navigator.rb
@@ -22,14 +22,13 @@ module Html2rss
     class << self
       ##
       # Extracts visible text from a given node and its children.
-      # Delegates to TextExtractor.
       #
       # @param tag [Nokogiri::XML::Node] the node from which to extract visible text
       # @param separator [String] separator used to join text fragments (default is a space)
       # @param exclude_nodes [Array<Nokogiri::XML::Node>, nil] nodes to exclude from extraction
       # @return [String, nil] the concatenated visible text, or nil if none is found
       def extract_visible_text(tag, separator: ' ', exclude_nodes: nil)
-        HtmlExtractor::TextExtractor.call(tag, separator:, exclude_nodes:)
+        TextExtractor.call(tag, separator:, exclude_nodes:)
       end
 
       ##
